refactor: extract ai settings section
